refactor/Login,Register : Refactor user/admin LoginPage, RegisterPage
