--- src/components/Common/Toggle/toggle.module.scss
@@ -1,33 +1,0 @@
-@use "../../../styles/constants" as c;
-
-.wrapper {
-    position: relative;
-    display: block;
-    width: 36px;
-    height: 20px;
-    cursor: pointer;
-    background-color: c.$C_WHITE_40;
-    border-radius: 50px;
-
-    .button {
-        position: absolute;
-        top: 50%;
-        left: 2px;
-        transform: translateY(-50%); /* 가운데를 맞추기 위해 top과 함께 쓰임 */
-        transition: left 0.3s ease-in; /* 버튼의 부드러운 이동 */
-        width: 16px;
-        height: 16px;
-        border-radius: 50%;
-        background-color: c.$C_WHITE;
-    }
-
-    /* 체크박스가 체크된 경우 부모선택자를 선택함
-     * 참고 : 부모 선택자 & 소개 https://sass-lang.com/documentation/style-rules/parent-selector
-     * 참고 : :has() 기능 소개 https://developer.mozilla.org/en-US/docs/Web/CSS/:has
-     * 참고 : clac() 기능 소개 https://developer.mozilla.org/en-US/docs/Web/CSS/calc */
-    &:has(input:checked) {
-        .button {
-            left: calc(100% - 18px); /* 버튼을 오른쪽으로 옮김: 100% - (버튼 width값 + 여유 공간값(2px)) */
-        }
-    }
-}
