+++ src/components/Common/index.js
@@ -7,7 +7,6 @@ export { default as Table } from "./Table";
 export { default as Tag } from "./Tag";
 export { default as Toast } from "./Toast";
 export { default as Stars } from "./Stars";
-export { default as Toggle } from "./Toggle";
 export { default as Card } from "./Card";
 export { default as Carousel } from "./Carousel";
 export { default as Paging } from "./Pagination";
