+++ src/pages/Auth/AdminLogin/adminLogin.module.scss
@@ -25,6 +25,7 @@
 
     > img {
       margin-bottom: 60px;
+      cursor: pointer;
     }
 
     > form {
