--- src/pages/Component/components.module.scss
@@ -1,55 +1,0 @@
-@use "../../styles/constants" as c;
-@use "../../styles/mixins/" as m;
-
-.wrapper {
-  overflow: scroll;
-  max-height: calc(100vh - 70px);
-  padding: 120px;
-  color: c.$C_WHITE;
-
-  > header {
-    margin-bottom: 80px;
-    font-size: 36px;
-    font-weight: 800;
-  }
-
-  > article {
-    @include m.flex(flex-start, flex-start, column);
-    gap: 48px;
-    > div {
-      > h2 {
-        margin-bottom: 8px;
-        font-size: 24px;
-        font-weight: 600;
-      }
-
-      > .buttons {
-        @include m.flex(flex-start);
-      }
-
-      > .tags {
-        @include m.flex(flex-start);
-        margin: 14px;
-      }
-      
-      > .title {
-                @include m.flex(space-between);
-
-                > h3 {
-                    margin-left: 22px;
-                    font-size: 20px;
-                    font-weight: 700;
-                }
-
-                > p {
-                    margin-right: 22px;
-                    font-size: 14px;
-                }
-      }
-
-      > .buttons {
-          @include m.flex(flex-start);
-      }
-    }
-  }
-}
